Auth/requirements.txt
daphne==4.1.2
django-cors-headers==4.5.0
pillow==11.0.0
psycopg==3.1.19
psycopg2-binary==2.9.9
channels==4.1.0
dj_rest_auth==6.0.0
Django==5.0.6
django_allauth==0.61.1
django_otp==1.5.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
PyJWT==2.10.0
pyotp==2.9.0
Requests==2.32.3
shortuuid==1.0.13
